avfilter/tonemapx: add dovi to hdr10 support
This adds a reshape-only mode for Dolby Vision videos without
a compatibility layer. In this mode, only Dolby Vision reshaping
will be performed, and the output will still be in SMPTE 2084
transfer.

The GPU-based filters already support this mode. This will be
useful in the future when we implement HDR transcoding.