avfilter/tonemap_opencl: implement tradeoff mode as 3dlut
This reworks the current tradeoff mode, which used to be a single
1D LUT for linearization, which is not fast enough on slow GPUs.
Instead, this makes the entire tonemap process a 3D LUT lookup.
This implementation first generates a 65x65x65 LUT on the GPU,
which has a similar compute cost to a single 400p frame. Then, it
uses tetrahedral interpolation to apply the LUT to the actual frame.
The interpolation quality is quite decent, and the interpolation
errors are very hard to notice unless in extreme conditions.

Signed-off-by: gnattu <gnattuoc@me.com>