avfilter/tonmap_opencl: handle overbrightness correctly in 3dlut
The old capping will use the second brightest cube instead of the
brightest one. The new capping will always use the brightest cube
and the values will be cancelled out during calculation, and only
the raw value of the brightest corner will left.

This is slightly slower as it introdced two more scalar ops, but
should still be fast enough.