lavc/videotoolbox: don't return external error for invalid vt frame
Some video sources are not reliable and will contain frames that
is not decodable with hardware accelerators. Current ffmpeg's
scheduling has low tolarance for decoder errors it does not understand.

Just return 0 after unref the frame instead so that ffmpeg will no
longer abort early due to videotoolbox is not able to decode a frame.